+++ src/IMSPublicApi/ReceivingOrderEndpoints/ROUpdateItemRequest.cs
@@ -3,7 +3,7 @@
     public class ROUpdateItemRequest : BaseRequest
     {
         public string ProductVariantId { get; set; }
-        public float QuantityUpdate { get; set; }
+        public int QuantityUpdate { get; set; }
         public string UnitUpdate { get; set; }
         public string CurrentReceivingOrderId { get; set; }
